(Currently Broken) Support using "count" api
At the moment this is broken. After some debugging I am pretty sure this is because the mz_at_ssl-Dsid cookie isn't valid for whatever reason. This needs to be looked into, but I have given up (and will probably just brute force all the years when looking up purchases).